# Improve performance of encoding