--- a/src/lib.rs
+++ b/src/lib.rs
@@ -172,16 +172,46 @@ pub trait FromHex: Sized {
     fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error>;
 }
 
-const fn val(c: u8, idx: usize) -> Result<u8, FromHexError> {
-    match c {
-        b'A'..=b'F' => Ok(c - b'A' + 10),
-        b'a'..=b'f' => Ok(c - b'a' + 10),
-        b'0'..=b'9' => Ok(c - b'0'),
-        _ => Err(FromHexError::InvalidHexCharacter {
-            c: c as char,
+const __: u8 = u8::MAX;
+
+// Lookup table for ascii to hex decoding.
+#[rustfmt::skip]
+static DECODE_TABLE: [u8; 256] = [
+    //   1   2   3   4   5   6   7   8   9   a   b   c   d   e   f
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 0
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 1
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 2
+     0,  1,  2,  3,  4,  5,  6,  7,  8,  9, __, __, __, __, __, __, // 3
+    __, 10, 11, 12, 13, 14, 15, __, __, __, __, __, __, __, __, __, // 4
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 5
+    __, 10, 11, 12, 13, 14, 15, __, __, __, __, __, __, __, __, __, // 6
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 7
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 8
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // 9
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // a
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // b
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // c
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // d
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // e
+    __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, __, // f
+];
+
+fn val(bytes: &[u8], idx: usize) -> Result<u8, FromHexError> {
+    let upper = DECODE_TABLE[bytes[0] as usize];
+    let lower = DECODE_TABLE[bytes[1] as usize];
+    if upper == u8::MAX {
+        return Err(FromHexError::InvalidHexCharacter {
+            c: bytes[0] as char,
             index: idx,
-        }),
+        });
     }
+    if lower == u8::MAX {
+        return Err(FromHexError::InvalidHexCharacter {
+            c: bytes[1] as char,
+            index: idx + 1,
+        });
+    }
+    Ok((upper << 4) | lower)
 }
 
 #[cfg(feature = "alloc")]
@@ -196,7 +226,7 @@ impl FromHex for Vec<u8> {
 
         hex.chunks(2)
             .enumerate()
-            .map(|(i, pair)| Ok(val(pair[0], 2 * i)? << 4 | val(pair[1], 2 * i + 1)?))
+            .map(|(i, pair)| val(pair, 2 * i))
             .collect()
     }
 }
@@ -319,8 +349,8 @@ pub fn decode_to_slice<T: AsRef<[u8]>>(data: T, out: &mut [u8]) -> Result<(), Fr
         return Err(FromHexError::InvalidStringLength);
     }
 
-    for (i, byte) in out.iter_mut().enumerate() {
-        *byte = val(data[2 * i], 2 * i)? << 4 | val(data[2 * i + 1], 2 * i + 1)?;
+    for (i, (data, byte)) in data.chunks(2).zip(out).enumerate() {
+        *byte = val(data, 2 * i)?;
     }
 
     Ok(())
@@ -340,7 +370,7 @@ fn generate_iter(len: usize) -> impl Iterator<Item = (usize, usize)> {
 // the inverse of `val`.
 #[inline]
 #[must_use]
-const fn byte2hex(byte: u8, table: &[u8; 16]) -> (u8, u8) {
+fn byte2hex(byte: u8, table: &[u8; 16]) -> (u8, u8) {
     let high = table[((byte & 0xf0) >> 4) as usize];
     let low = table[(byte & 0x0f) as usize];